Modify KVMessage to match assignment specs
Previous implementation of the new server metadata information features
for the client side erroneously assumed that metadata for all servers
was to be passed in a SERVER_NOT_RESPONSIBLE message. However, the
assignment stub defines the interface as providing only a single
server's information (the one responsible for the key in the request).
Modify client and serialization classes to accommodate this.

Another realization was that there was an assignment-defined interface
for server metadata (IECSNode). Make slight modifications to
ServerMetadata in order to serve as an implementation for this
interface.