Fix issue where logs would be created relative to root
Because of the way the server initialization script is called, all file
paths in the server application are created relative to the root
directory. Add way for server to find where the correct log directory is
supposed to be.

Also remove custom pattern for console appender, since there is no
longer any convenient way to access it.